Refactors SVE functions to separate header file
