+++ Code/Numerics/Optimizer/CMakeLists.txt
@@ -1,10 +1,10 @@
 
-rdkit_library(Optimizer 
+rdkit_library(Optimizer
               BFGSOpt.cpp LinearSearch.cpp
               LINK_LIBRARIES RDGeometryLib Trajectory RDGeneral)
 target_compile_definitions(Optimizer PRIVATE RDKIT_OPTIMIZER_BUILD)
 
-rdkit_headers(BFGSOpt.h DEST Numerics/Optimizer)
+rdkit_headers(BFGSOpt.h BFGSOpt_SVE.h DEST Numerics/Optimizer)
 
 rdkit_catch_test(testOptimizer testOptimizer.cpp LINK_LIBRARIES Optimizer )
